requirements.txt
dbt-bigquery
pandas
pyspark
google-cloud-bigquery
google-cloud-storage
requests
pytest
ydata-profiling
python-dotenv>=1.0.1
jupyterlab>=1.0.0